tratamento de erros terminado no servidor de arquivos

+++ src/files/Controller/ImageController.js
@@ -17,30 +17,73 @@ class ImageController{
     }
     static async Create(req,res) {
         const {base64,path}=req.body
+        if(!base64||!path){
+            return res.status(400).json({"sucess":false,"message":"Parâmetros inválidos"})
+        }
         const Separated=ImageController.Separate(path)
+        const Arquivo_sepado=Separated[0].split(".")
+        if(Arquivo_sepado.length!=2||Arquivo_sepado[1]==="."){
+            return res.status(400).json({"sucess":false,"message":"Nome de arquivo inválido"})
+        }
         const DadFolderPath=Separated[1]
         console.log(DadFolderPath)
         const folder=new Folder(DadFolderPath)
-        await folder.Create()
-        await Image.salvar(path,base64)
-        return res.json("Arquivo salvo com sucesso")
+        
+        try{
+            if(! await folder.pastaExiste()){
+                await folder.Create()
+            }
+            await Image.salvar(path,base64)
+            return res.status(200).json({"sucess":true,"message":"arquivo salvo com sucesso"})
+        }catch(error){
+            console.error(error)
+            return res.status(501).json({"sucess":false,"message":"Erro interno do servidor"})
+        }
     }
     static async Read(req,res){
         var {path} = req.body
+        if(!path){
+            return res.status(400).json({"sucess":false,"message":"Parâmetros inválidos"})
+        }
         const Separated=ImageController.Separate(path)
         const name=Separated[0]
-        const DadFolderPath=Separated[1]
-        const imagem=new Image(DadFolderPath,name)
-        return res.json(await imagem.read())
+        try{
+            const DadFolderPath=new Folder(Separated[1])
+            const imagem=new Image(Separated[1],name)
+            if(! await DadFolderPath.pastaExiste() || ! await DadFolderPath.FileExist(imagem)){
+                return res.status(400).json({"sucess":false,"message":"Arquivo inexistente"})
+            }
+            return res.status(200).json({"sucess":true,"message":"Leitura realizada com sucesso","data":await imagem.read()})
+        }
+        catch(error){
+            console.error(error)
+            return res.status(501).json({"sucess":false,"message":"Erro interno do servidor"})
+        }
     }
     static async Delete(req,res){
         var {path} = req.body
-        const Desestruturado=ImageController.Separate(path)
-        const name=Desestruturado[0]
-        const Dad=Desestruturado[1]
+        if(!path){
+            return res.status(400).json({"sucess":false,"message":"Parâmetros inválidos"})   
+        }
+        const Separated=ImageController.Separate(path)
+        const name=Separated[0]
+        const Dad=Separated[1]
         const imagem=new Image(Dad,name)
-        await imagem.delete()
-        return res.json("arquivo deletado com sucesso")
+        try{
+            const DadFolderPath=new Folder(Separated[1])
+            if(! await DadFolderPath.pastaExiste() || ! await DadFolderPath.FileExist(imagem)){
+                return res.status(400).json({"sucess":false,"message":"Arquivo inexistente"})
+            }
+            await imagem.delete()
+            return res.json("arquivo deletado com sucesso")
+        }catch(error){
+            if (error.code === 'ENOENT') {
+                return res.status(401).json({"sucess":false,"message":"Arquivo inexistente"})
+            }
+            console.error(error)
+            return res.status(501).json({"sucess":false,"message":"Erro interno do servidor"})
+        }
+
     }
 }
 export default ImageController
